+++ src/bindings.c
@@ -18,7 +18,7 @@ static int tolua_levitate_ship(lua_State * L)
 {
   ship *sh = (ship *) tolua_tousertype(L, 1, 0);
   unit *mage = (unit *) tolua_tousertype(L, 2, 0);
-  double power = (double)tolua_tonumber(L, 3, 0);
+  float power = (float)tolua_tonumber(L, 3, 0);
   int duration = (int)tolua_tonumber(L, 4, 0);
   int cno = levitate_ship(sh, mage, power, duration);
   tolua_pushnumber(L, (lua_Number) cno);
