+++ src/spells/shipcurse.h
@@ -26,8 +26,8 @@ extern "C" {
     const struct curse *c, int self);
   extern void register_shipcurse(void);
   extern struct curse *shipcurse_flyingship(struct ship *sh, struct unit *mage,
-    double power, int duration);
-  int levitate_ship(struct ship *sh, struct unit *mage, double power,
+    float power, int duration);
+  int levitate_ship(struct ship *sh, struct unit *mage, float power,
     int duration);
 
 #ifdef __cplusplus
